requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile
#
backoff==1.11.1
    # via -r requirements.in
boto3==1.18.43
    # via -r requirements.in
botocore==1.21.43
    # via
    #   boto3
    #   s3transfer
certifi==2021.5.30
    # via requests
charset-normalizer==2.0.5
    # via requests
click==8.0.1
    # via flask
dnspython==2.1.0
    # via -r requirements.in
flask==2.0.1
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.2
    # via requests
importlib-metadata==4.8.1
    # via click
itsdangerous==2.0.1
    # via flask
jinja2==3.0.1
    # via flask
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
markupsafe==2.0.1
    # via jinja2
python-dateutil==2.8.2
    # via botocore
requests==2.26.0
    # via -r requirements.in
s3transfer==0.5.0
    # via boto3
six==1.16.0
    # via python-dateutil
typing-extensions==3.10.0.2
    # via importlib-metadata
urllib3==1.26.6
    # via
    #   botocore
    #   requests
werkzeug==2.0.1
    # via flask
zipp==3.5.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
